fix the pull request issues
