--- Physical/ImagePicker.swift
@@ -1,51 +1,0 @@
-import SwiftUI
-
-public struct CameraPickerView: UIViewControllerRepresentable {
-    
-    private var sourceType: UIImagePickerController.SourceType = .camera
-    private let onImagePicked: (UIImage) -> Void
-    
-    @Environment(\.presentationMode) private var presentationMode
-    
-    public init(onImagePicked: @escaping (UIImage) -> Void) {
-        self.onImagePicked = onImagePicked
-    }
-    
-    public func makeUIViewController(context: Context) -> UIImagePickerController {
-        let picker = UIImagePickerController()
-        picker.sourceType = self.sourceType
-        picker.delegate = context.coordinator
-        return picker
-    }
-    
-    public func updateUIViewController(_ uiViewController: UIImagePickerController, context: Context) {}
-    
-    public func makeCoordinator() -> Coordinator {
-        Coordinator(
-            onDismiss: { self.presentationMode.wrappedValue.dismiss() },
-            onImagePicked: self.onImagePicked
-        )
-    }
-    
-    final public class Coordinator: NSObject, UINavigationControllerDelegate, UIImagePickerControllerDelegate {
-        
-        private let onDismiss: () -> Void
-        private let onImagePicked: (UIImage) -> Void
-        
-        init(onDismiss: @escaping () -> Void, onImagePicked: @escaping (UIImage) -> Void) {
-            self.onDismiss = onDismiss
-            self.onImagePicked = onImagePicked
-        }
-        
-        public func imagePickerController(_ picker: UIImagePickerController,
-                                          didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey: Any]) {
-            if let image = info[.originalImage] as? UIImage {
-                self.onImagePicked(image)
-            }
-            self.onDismiss()
-        }
-        public func imagePickerControllerDidCancel(_: UIImagePickerController) {
-            self.onDismiss()
-        }
-    }
-}
