+++ TikiClient/Classes/Capture/CaptureService.swift
@@ -1,32 +1,17 @@
-import UIKit
 
 /// Service for capturing and processing receipt data.
-public class CaptureService: UIViewController, UIImagePickerControllerDelegate, UINavigationControllerDelegate {
-
-    var imagePickerCallback: ((UIImage) -> Void)? = nil
-    var imagePickerRootVc: UIViewController? = nil
+public class CaptureService {
+    
+    var captureScan = CaptureScan()
 
     /// Captures an image of a receipt for processing.
     ///
     /// - Returns: The captured receipt image.
     public func scan(onFinish: @escaping (UIImage) -> Void){
-      imagePickerCallback = onFinish
-      let keyWindow = UIApplication.shared.windows.first(where: { $0.isKeyWindow }) ?? UIApplication.shared.windows.first
-      imagePickerRootVc = keyWindow?.rootViewController
-      let imagePickerController = UIImagePickerController()
-      imagePickerController.allowsEditing = false
-      imagePickerController.sourceType = .camera
-      imagePickerController.delegate = self
-      imagePickerRootVc?.present(imagePickerController, animated: true, completion: nil)
+        captureScan.captureScan(onFinish: onFinish)
     }
 
-    public func imagePickerController(_ picker: UIImagePickerController, didFinishPickingMediaWithInfo info: [UIImagePickerController.InfoKey : Any]) {
-        let tempImage = info[UIImagePickerController.InfoKey.originalImage] as! UIImage
-        imagePickerCallback?(tempImage);
-        imagePickerRootVc?.dismiss(animated: true, completion: nil)
-        imagePickerCallback = nil
-        imagePickerRootVc = nil
-    }
+
 
     /// Downloads potential receipt data from known receipt email senders and publishes it.
     ///
