--- TikiClient/Classes/Capture/PublishingStatusEnum.swift
@@ -1,12 +1,0 @@
-import Foundation
-
-public enum PublishingStatusEnum: String {
-    /// In-progress status.
-    case inProgress = "IN_PROGRESS"
-    
-    /// Error status.
-    case error = "ERROR"
-    
-    /// Done status.
-    case done = "DONE"
-}
